pull out code sample and block into separate components
